+++ app/gui/src/project-view/components/GraphEditor.vue
@@ -17,6 +17,7 @@ import GraphNodes from '@/components/GraphEditor/GraphNodes.vue'
 import { useGraphEditorClipboard } from '@/components/GraphEditor/clipboard'
 import { performCollapse, prepareCollapsedInfo } from '@/components/GraphEditor/collapsing'
 import type { NodeCreationOptions } from '@/components/GraphEditor/nodeCreation'
+import { registerSelectionActionHandlers } from '@/components/GraphEditor/selectionActions'
 import { useGraphEditorToasts } from '@/components/GraphEditor/toasts'
 import { uploadedExpression, Uploader } from '@/components/GraphEditor/upload'
 import GraphMissingView from '@/components/GraphMissingView.vue'
@@ -30,7 +31,6 @@ import { useDoubleClick } from '@/composables/doubleClick'
 import { keyboardBusy, keyboardBusyExceptIn, unrefElement, useEvent } from '@/composables/events'
 import { groupColorVar } from '@/composables/nodeColors'
 import type { PlacementStrategy } from '@/composables/nodeCreation'
-import { provideGraphEditorLayers } from '@/providers/graphEditorLayers'
 import { provideGraphEditorState } from '@/providers/graphEditorState'
 import type { GraphNavigator } from '@/providers/graphNavigator'
 import { provideGraphNavigator } from '@/providers/graphNavigator'
@@ -40,7 +40,6 @@ import { provideGraphSelection } from '@/providers/graphSelection'
 import { provideStackNavigator } from '@/providers/graphStackNavigator'
 import { provideInteractionHandler } from '@/providers/interactionHandler'
 import { provideKeyboard } from '@/providers/keyboard'
-import { provideSelectionButtons } from '@/providers/selectionButtons'
 import { provideWidgetRegistry } from '@/providers/widgetRegistry'
 import type { Node, NodeId } from '@/stores/graph'
 import { provideGraphStore } from '@/stores/graph'
@@ -103,15 +102,6 @@ const viewportNode = ref<HTMLElement>()
 onMounted(() => viewportNode.value?.focus())
 const graphNavigator: GraphNavigator = provideGraphNavigator(viewportNode, keyboard, {
   predicate: (e) => (e instanceof KeyboardEvent ? nodeSelection.selected.size === 0 : true),
-})
-
-// === Exposed layers ===
-
-const rootNode = ref<HTMLElement>()
-const floatingLayer = ref<HTMLElement>()
-provideGraphEditorLayers({
-  fullscreen: rootNode,
-  floating: floatingLayer,
 })
 
 // === Client saved state ===
@@ -230,7 +220,7 @@ const { copyNodesToClipboard, createNodesFromClipboard } = useGraphEditorClipboa
 
 // === Selection Buttons ===
 
-const { buttons: selectionButtons } = provideSelectionButtons(
+const selectionHandlers = registerSelectionActionHandlers(
   () =>
     iter.filterDefined(
       iter.map(
@@ -295,7 +285,7 @@ const graphBindingsHandler = graphBindings.handler({
       createWithComponentBrowser(fromSelection() ?? { placement: { type: 'mouse' } })
     }
   },
-  deleteSelected: selectionButtons.deleteSelected.action!,
+  deleteSelected: selectionHandlers['components.deleteSelected'].action,
   zoomToSelected() {
     zoomToSelected()
   },
@@ -319,11 +309,11 @@ const graphBindingsHandler = graphBindings.handler({
       }
     })
   },
-  copyNode: selectionButtons.copy.action!,
+  copyNode: selectionHandlers['components.copy'].action,
   pasteNode() {
     createNodesFromClipboard()
   },
-  collapse: selectionButtons.collapse.action!,
+  collapse: selectionHandlers['components.collapse'].action,
   enterNode() {
     const selectedNode = set.first(nodeSelection.selected)
     if (selectedNode) {
@@ -334,7 +324,7 @@ const graphBindingsHandler = graphBindings.handler({
     stackNavigator.exitNode()
   },
   changeColorSelectedNodes() {
-    selectionButtons.pickColorMulti.state = true
+    selectionHandlers['components.pickColorMulti'].toggled.value = true
   },
   openDocumentation() {
     const result = tryGetSelectionDocUrl()
@@ -628,7 +618,7 @@ const groupColors = computed(() => {
 
 <template>
   <div
-    ref="rootNode"
+    id="graphEditorRoot"
     class="GraphEditor"
     :class="{ draggingEdge: graphStore.mouseEditedEdge != null }"
     :style="groupColors"
@@ -681,11 +671,6 @@ const groupColors = computed(() => {
           :scrollableArea="Rect.Bounding(...graphStore.visibleNodeAreas)"
         />
         <GraphMouse />
-        <div
-          ref="floatingLayer"
-          class="floatingLayer"
-          :style="{ transform: graphNavigator.transform }"
-        />
       </div>
       <BottomPanel v-model:show="showCodeEditor">
         <Suspense>
@@ -736,20 +721,5 @@ const groupColors = computed(() => {
   touch-action: none;
   --node-color-no-type: #596b81;
   --output-node-color: #006b8a;
-}
-
-.floatingLayer {
-  position: absolute;
-  top: 0;
-  left: 0;
-  /* The size isn't important, except it must be non-zero for `floating-ui` to calculate the scale factor. */
-  width: 1px;
-  height: 1px;
-  contain: layout size style;
-  will-change: transform;
-  pointer-events: none;
-  > * {
-    pointer-events: auto;
-  }
 }
 </style>