+++ app/gui/src/project-view/components/MarkdownEditor/codemirror/decoration/index.ts
@@ -1,0 +1,26 @@
+import { cursorDecoratorExt } from '@/components/MarkdownEditor/codemirror/decoration/editingAtCursor'
+import {
+  decorateImageWithClass,
+  decorateImageWithRendered,
+  linkDecoratorExt,
+  markdownLinkEditPopup,
+} from '@/components/MarkdownEditor/codemirror/decoration/linksAndImages'
+import { listDecoratorExt } from '@/components/MarkdownEditor/codemirror/decoration/lists'
+import { decorateTable } from '@/components/MarkdownEditor/codemirror/decoration/table'
+import { treeStateDecorator } from '@/components/MarkdownEditor/codemirror/decoration/treeStateDecorator'
+import { linkDecoratorStateExt } from '@/util/codemirror/links'
+import { vueHostExt } from '@/util/codemirror/vueHostExt'
+import { type Extension } from '@codemirror/state'
+
+/** Extension applying decorators for Markdown. */
+export function markdownDecorators(): Extension {
+  return [
+    linkDecoratorStateExt,
+    vueHostExt,
+    treeStateDecorator([decorateImageWithClass, decorateImageWithRendered, decorateTable]),
+    linkDecoratorExt(),
+    listDecoratorExt(),
+    cursorDecoratorExt(),
+    markdownLinkEditPopup(),
+  ]
+}