+++ app/gui/src/project-view/components/widgets/FileBrowserWidget.vue
@@ -5,10 +5,14 @@ export default {
 </script>
 
 <script setup lang="ts">
+import ContextMenuTrigger from '@/components/ContextMenuTrigger.vue'
 import LoadingSpinner from '@/components/shared/LoadingSpinner.vue'
 import SvgButton from '@/components/SvgButton.vue'
 import SvgIcon from '@/components/SvgIcon.vue'
+import FileBrowserEntry from '@/components/widgets/FileBrowserWidget/FileBrowserEntry.vue'
+import { Directory, useFileBrowserStack } from '@/components/widgets/FileBrowserWidget/paths'
 import { useBackend } from '@/composables/backend'
+import { Action } from '@/providers/action'
 import { injectBackend } from '@/providers/backend'
 import { assert } from '@/util/assert'
 import type { ToValue } from '@/util/reactivity'
@@ -24,11 +28,15 @@ import Backend, {
   assetIsDirectory,
   assetIsFile,
 } from 'enso-common/src/services/Backend'
-import { computed, onMounted, reactive, ref, toValue, watch } from 'vue'
-import { Err, Ok, Result } from 'ydoc-shared/util/data/result'
-import FileBrowserEntry from './FileBrowserWidget/FileBrowserEntry.vue'
+import { computed, onMounted, reactive, ref, toRef, toValue, watch } from 'vue'
 
-const { writeMode = false } = defineProps<{ writeMode?: boolean }>()
+const props = withDefaults(
+  defineProps<{
+    writeMode?: boolean
+    choosenPath?: string
+  }>(),
+  { writeMode: false, choosenPath: '' },
+)
 
 const emit = defineEmits<{
   pathAccepted: [path: string]
@@ -38,8 +46,8 @@ const { query, fetch, ensureQueryData, mutation } = useBackend('remote')
 const { remote: backend } = injectBackend()
 
 const errorToast = useToast.error()
-const fileName = ref<string>('')
 const newDirPlaceholder = Symbol()
+
 let nextKeyForNewDir = 0
 /**
  * Override for `:key` attribute in content entries.
@@ -50,28 +58,24 @@ let nextKeyForNewDir = 0
  */
 const keyOverride: Map<DirectoryId | symbol, number> = reactive(new Map())
 
-// === Current Directory ===
-
-interface Directory {
-  id: DirectoryId
-  title: string
-}
-
 const currentUser = query('usersMe', [])
 const currentOrganization = query('getOrganization', [])
-const directoryStack = ref<Directory[]>([])
-const isDirectoryStackInitializing = computed(() => directoryStack.value.length === 0)
-const currentDirectory = computed(() => directoryStack.value[directoryStack.value.length - 1])
 
-const currentPath = computed(() => {
-  if (!currentUser.data.value) return
-  let root = backend?.rootPath(currentUser.data.value) ?? 'enso://'
-  if (!root.endsWith('/')) root += '/'
-  return `${root}${directoryStack.value
-    .slice(1)
-    .map((dir) => `${dir.title}/`)
-    .join('')}`
-})
+const {
+  filenameInputContents,
+  directoryStack,
+  currentDirectory,
+  currentFilePath,
+  highlightedName,
+  initializeStack,
+  isDirectoryStackInitializing,
+} = useFileBrowserStack(
+  backend,
+  toRef(props, 'choosenPath'),
+  currentUser.data,
+  toRef(props, 'writeMode'),
+  (dir) => fetch('listDirectory', listDirectoryArgs(dir)),
+)
 
 // === Directory Contents ===
 
@@ -125,21 +129,13 @@ function enterDir(dir: DirectoryAsset) {
   directoryStack.value.push(dir)
 }
 
-class DirNotFoundError {
-  constructor(public dirName: string) {}
-
-  toString() {
-    return `Directory "${this.dirName}" not found`
-  }
-}
-
 function popTo(index: number) {
   directoryStack.value.splice(index + 1)
 }
 
 function chooseFile(file: FileAsset | DatalinkAsset) {
-  fileName.value = file.title
-  if (!writeMode) {
+  filenameInputContents.value = file.title
+  if (!props.writeMode) {
     acceptCurrentFile()
   }
 }
@@ -157,31 +153,36 @@ const isBusy = computed(() => isDirectoryStackInitializing.value || isPending.va
 const anyError = computed(() =>
   isError.value ? error
   : currentUser.isError.value ? currentUser.error
+  : currentOrganization.isError.value ? currentOrganization.error
   : undefined,
-)
-
-const currentFilePath = computed(
-  () => fileName.value && currentPath.value && `${currentPath.value}${fileName.value}`,
 )
 
 // === Creating and Renaming Directories ===
 
 const editedAsset = ref<{
-  asset: Directory | typeof newDirPlaceholder
+  asset?: Directory
   name: string
   state: 'editing' | 'pending' | 'just created'
   createdId?: DirectoryId
 }>()
+
+// Don't await invalidates, because we want `createDirectory` to return first, to fill
+// `keyOverride` property before getting update from backend.
 const createDir = mutation('createDirectory', { meta: { awaitInvalidates: false } })
 const updateDir = mutation('updateDirectory')
 
 function addNewDirectory() {
   assert(editedAsset.value == null)
   keyOverride.set(newDirPlaceholder, nextKeyForNewDir++)
-  editedAsset.value = { asset: newDirPlaceholder, name: 'New Folder', state: 'editing' }
+  editedAsset.value = { name: 'New Folder', state: 'editing' }
 }
 
-function acceptName(name: string, actionDescription: string) {
+function renameDirectory(dir: DirectoryAsset) {
+  assert(editedAsset.value == null)
+  editedAsset.value = { asset: dir, name: dir.title, state: 'editing' }
+}
+
+async function acceptName(name: string) {
   if (editedAsset.value?.state !== 'editing') {
     console.error('Accepting edited name without editing')
     return
@@ -194,68 +195,60 @@ function acceptName(name: string, actionDescription: string) {
     console.error('Cannot rename directory without parentId')
     return
   }
-  const requestBody = { title: edited.name, parentId }
   const action =
-    edited.asset === newDirPlaceholder ?
-      createDir.mutateAsync([requestBody, false])
-    : updateDir.mutateAsync([edited.asset.id, requestBody, edited.asset.title])
-  action
-    .then((result) => {
+    edited.asset == null ? createDir.mutateAsync([{ title: edited.name, parentId }, false])
+    : edited.asset.title != edited.name ?
+      updateDir.mutateAsync([edited.asset.id, { title: edited.name }, edited.asset.title])
+    : Promise.resolve(undefined)
+  action.then(
+    (result) => {
       assert(edited === editedAsset.value)
-      if (result?.id) {
-        editedAsset.value.createdId = result.id
-        editedAsset.value.state = 'just created'
+      // Editing existing asset does not require 'just created' state, because we await
+      // invalidates there
+      if (edited.asset == null && result != null) {
+        edited.createdId = result.id
+        edited.state = 'just created'
         const key = keyOverride.get(newDirPlaceholder)
         if (key != null) {
           keyOverride.set(result.id, key)
         }
+      } else {
+        editedAsset.value = undefined
       }
-    })
-    .catch((error) => {
+    },
+    (error) => {
+      const actionDescription = edited.asset == null ? 'create folder' : 'rename folder'
       errorToast.show(`Failed to ${actionDescription}: ${error}`)
       editedAsset.value = undefined
-    })
+    },
+  )
 }
 
 watch(
   directories,
   (dirs) => {
-    // Remove placeholder once received an actual directory.
-    if (dirs?.find((dir) => dir.id === editedAsset.value?.createdId)) editedAsset.value = undefined
+    // Finish editing once received an updated directory.
+    if (dirs?.find((dir) => dir.id === editedAsset.value?.createdId)) {
+      editedAsset.value = undefined
+    }
   },
   { flush: 'sync' },
 )
+// Currently, the only way to "focus" on an element is by context menu.
+const focusedDirectory = ref<DirectoryAsset>()
+const renameAction: Action = {
+  icon: 'edit',
+  description: 'Rename directory',
+  disabled: computed(() => focusedDirectory.value == null || editedAsset.value != null),
+  action: () => focusedDirectory.value && renameDirectory(focusedDirectory.value),
+}
 
 // === Initialization ===
 
-async function enterDirByName(name: string, stack: Directory[]): Promise<Result> {
-  const currentDir = stack[stack.length - 1]
-  if (currentDir == null) return Err('Stack is empty')
-  const content = await fetch('listDirectory', listDirectoryArgs(currentDir))
-  const nextDir = content.find(
-    (asset): asset is DirectoryAsset => assetIsDirectory(asset) && asset.title === name,
-  )
-  if (!nextDir) return Err(new DirNotFoundError(name))
-  stack.push(nextDir)
-  return Ok()
-}
-
 onMounted(() => {
   Promise.all([currentUser.promise.value, currentOrganization.promise.value]).then(
-    async ([user, organization]) => {
-      if (!user) {
-        errorToast.show('Cannot load file list: not logged in.')
-        return
-      }
-      const rootDirectoryId =
-        backend?.rootDirectoryId(user, organization, null) ?? user.rootDirectoryId
-      const stack = [{ id: rootDirectoryId, title: 'Cloud' }]
-      if (rootDirectoryId != user.rootDirectoryId) {
-        let result = await enterDirByName('Users', stack)
-        result = result.ok ? await enterDirByName(user.name, stack) : result
-        if (!result.ok) errorToast.reportError(result.error, 'Cannot enter home directory')
-      }
-      directoryStack.value = stack
+    ([user, organizaton]) => {
+      initializeStack(user, organizaton)
     },
   )
 })
@@ -285,40 +278,44 @@ onMounted(() => {
       />
     </div>
 
-    <div v-if="isBusy" class="centerContent contents"><LoadingSpinner /></div>
-    <div v-else-if="anyError" class="centerContent contents">Error: {{ anyError }}</div>
+    <div v-if="anyError" class="centerContent contents">Error: {{ anyError }}</div>
+    <div v-else-if="isBusy" class="centerContent contents"><LoadingSpinner /></div>
     <div v-else-if="isEmpty" class="centerContent contents">Directory is empty</div>
     <div v-else :key="currentDirectory?.id ?? 'root'" class="listing contents">
-      <TransitionGroup>
-        <FileBrowserEntry
-          v-if="editedAsset?.asset === newDirPlaceholder"
-          :key="keyOverride.get(newDirPlaceholder) ?? newDirPlaceholder"
-          icon="folder"
-          :title="editedAsset.name"
-          :editingState="editedAsset.state"
-          @nameAccepted="acceptName($event, 'create folder')"
-        />
-        <FileBrowserEntry
-          v-for="entry in directories"
-          :key="keyOverride.get(entry.id) ?? entry.id"
-          icon="folder"
-          :title="editedAsset?.asset === entry ? editedAsset.name : entry.title"
-          :editingState="editedAsset?.asset === entry ? editedAsset.state : undefined"
-          @click="enterDir(entry)"
-          @nameAccepted="acceptName($event, 'rename folder')"
-        />
-        <FileBrowserEntry
-          v-for="entry in files"
-          :key="entry.id"
-          icon="text2"
-          :title="entry.title"
-          @click="chooseFile(entry)"
-        />
-      </TransitionGroup>
+      <ContextMenuTrigger :actions="[renameAction]" @hidden="focusedDirectory = undefined">
+        <TransitionGroup>
+          <FileBrowserEntry
+            v-if="editedAsset && editedAsset.asset == null"
+            :key="keyOverride.get(newDirPlaceholder) ?? newDirPlaceholder"
+            icon="folder"
+            :title="editedAsset.name"
+            :editingState="editedAsset.state"
+            @nameAccepted="acceptName($event)"
+          />
+          <FileBrowserEntry
+            v-for="entry in directories"
+            :key="keyOverride.get(entry.id) ?? entry.id"
+            icon="folder"
+            :title="editedAsset?.asset?.id === entry.id ? editedAsset.name : entry.title"
+            :editingState="editedAsset?.asset?.id === entry.id ? editedAsset.state : undefined"
+            @click="enterDir(entry)"
+            @nameAccepted="acceptName($event)"
+            @contextmenu="focusedDirectory = entry"
+          />
+          <FileBrowserEntry
+            v-for="entry in files"
+            :key="entry.id"
+            icon="text2"
+            :title="entry.title"
+            :highlighted="entry.title === highlightedName"
+            @click="chooseFile(entry)"
+          />
+        </TransitionGroup>
+      </ContextMenuTrigger>
     </div>
     <div v-if="writeMode" class="fileNameBar">
       <input
-        v-model="fileName"
+        v-model="filenameInputContents"
         class="fileNameInput"
         @pointerdown.stop
         @click.stop
@@ -332,7 +329,7 @@ onMounted(() => {
       <SvgButton
         class="fileNameAcceptButton"
         label="Ok"
-        :disabled="!fileName"
+        :disabled="!filenameInputContents"
         @click.stop="acceptCurrentFile"
       />
     </div>