+++ std-bits/table/src/main/java/org/enso/table/data/column/storage/StringStorage.java
@@ -10,10 +10,7 @@ import org.enso.table.data.column.operation.CountUntrimmed;
 import org.enso.table.data.column.operation.SampleOperation;
 import org.enso.table.data.column.operation.map.MapOperationStorage;
 import org.enso.table.data.column.operation.map.text.CoalescingStringStringOp;
-import org.enso.table.data.column.operation.map.text.LikeOp;
-import org.enso.table.data.column.operation.map.text.StringBooleanOp;
 import org.enso.table.data.column.operation.map.text.StringIsInOp;
-import org.enso.table.data.column.operation.map.text.StringLongToStringOp;
 import org.enso.table.data.column.operation.map.text.StringStringOp;
 import org.enso.table.data.column.storage.type.StorageType;
 import org.enso.table.data.column.storage.type.TextType;
@@ -25,7 +22,6 @@ public final class StringStorage extends SpecializedStorage<String> {
   private static final Logger LOGGER = org.slf4j.LoggerFactory.getLogger(StringStorage.class);
 
   record DataQualityMetrics(Long untrimmedCount, Long whitespaceCount) {}
-  ;
 
   private Future<DataQualityMetrics> dataQualityMetricsValues;
 
@@ -37,7 +33,7 @@ public final class StringStorage extends SpecializedStorage<String> {
     super(type, data, buildOps());
 
     dataQualityMetricsValues =
-        CompletableFuture.supplyAsync(() -> createDataQualityMetricsWitDefaultSize());
+        CompletableFuture.supplyAsync(this::createDataQualityMetricsWitDefaultSize);
   }
 
   public static StringStorage makeEmpty(TextType type, long size) {
@@ -111,42 +107,6 @@ public final class StringStorage extends SpecializedStorage<String> {
 
   private static MapOperationStorage<String, SpecializedStorage<String>> buildOps() {
     MapOperationStorage<String, SpecializedStorage<String>> t = new MapOperationStorage<>();
-    t.add(
-        new StringBooleanOp(Maps.STARTS_WITH) {
-          @Override
-          protected boolean doString(String a, String b) {
-            return Text_Utils.starts_with(a, b);
-          }
-        });
-    t.add(
-        new StringBooleanOp(Maps.ENDS_WITH) {
-          @Override
-          protected boolean doString(String a, String b) {
-            return Text_Utils.ends_with(a, b);
-          }
-        });
-    t.add(
-        new StringLongToStringOp(Maps.TEXT_LEFT) {
-          @Override
-          protected String doOperation(String a, long b) {
-            return Text_Utils.take_prefix(a, b);
-          }
-        });
-    t.add(
-        new StringLongToStringOp(Maps.TEXT_RIGHT) {
-          @Override
-          protected String doOperation(String a, long b) {
-            return Text_Utils.take_suffix(a, b);
-          }
-        });
-    t.add(
-        new StringBooleanOp(Maps.CONTAINS) {
-          @Override
-          protected boolean doString(String a, String b) {
-            return Text_Utils.contains(a, b);
-          }
-        });
-    t.add(new LikeOp());
     t.add(new StringIsInOp<>());
     t.add(
         new StringStringOp(Maps.ADD) {