+++ CHANGELOG.md
@@ -26,6 +26,13 @@
   (new) files.][12228]
 - [Cloud File Browser allows adding new directory in "writing"
   components][12275]
+- [In the documentation panel, text can now be made bold or italic using
+  buttons][12341]
+- [Cloud File Browser allows renaming existing directories in "writing"
+  components][12323]
+- ["Insert link" button added to documentation panel][12365]
+- [Cloud File Browser, when opened first time after opening project, shows and
+  highlights the currently set file][12184]
 
 [11889]: https://github.com/enso-org/enso/pull/11889
 [11836]: https://github.com/enso-org/enso/pull/11836
@@ -43,6 +50,10 @@
 [12228]: https://github.com/enso-org/enso/pull/12228
 [12217]: https://github.com/enso-org/enso/pull/12217
 [12275]: https://github.com/enso-org/enso/pull/12275
+[12341]: https://github.com/enso-org/enso/pull/12341
+[12323]: https://github.com/enso-org/enso/pull/12323
+[12365]: https://github.com/enso-org/enso/pull/12365
+[12184]: https://github.com/enso-org/enso/pull/12184
 
 #### Enso Standard Library
 
@@ -62,6 +73,8 @@
 - [Support for key-pair authentication in Snowflake connector.][12247]
 - [Support for basic arithmetic operations as numbers in Expressions.][12297]
 - [Support for Regular Expressions in Enso Expressions.][12320]
+- [Support for pi() and e() in Enso Expressions.][12367]
+- [xlsx reader now does not read empty rows from the end of a worksheet][12345]
 
 [11926]: https://github.com/enso-org/enso/pull/11926
 [12031]: https://github.com/enso-org/enso/pull/12031
@@ -74,6 +87,8 @@
 [12247]: https://github.com/enso-org/enso/pull/12247
 [12297]: https://github.com/enso-org/enso/pull/12297
 [12320]: https://github.com/enso-org/enso/pull/12320
+[12367]: https://github.com/enso-org/enso/pull/12367
+[12345]: https://github.com/enso-org/enso/pull/12345
 
 #### Enso Language & Runtime